refactor: 拆分 AwardGallery，拥有更清晰的层级
