+++ modern/src/App.js
@@ -1,20 +1,25 @@
 import React from 'react';
-import AwardGallery from './components/AwardGallery';
 import Footer from './components/Footer';
 import Header from './components/Header';
+import GalleryContainer from './components/GalleryContainer';
+import AwardGallery from './components/AwardGallery';
 import awards from './configs/awards';
 
 function App() {
   return (
     <div id="wrapper">
       <Header />
-      {awards.map(awardList => (
-        <AwardGallery
-          key={awardList.title}
-          title={awardList.title}
-          awards={awardList.awards}
-        />
-      ))}
+      <GalleryContainer>
+        <ul>
+          {awards.map(awardList => (
+            <AwardGallery
+              key={awardList.title}
+              title={awardList.title}
+              awards={awardList.awards}
+            />
+          ))}
+        </ul>
+      </GalleryContainer>
       <Footer />
     </div>
   );
